bot/summary.py
from typing import Optional, Dict, Any
from datetime import datetime, timedelta
from discord.ext import commands
from discord import TextChannel


class SummaryGenerator:
    def __init__(self, message_store, config):
        """Initialize summary generator with message store and config"""
        self.message_store = message_store
        self.config = config

    def generate_daily_summary(self) -> str:
        """
        Generate a daily summary of messages
        TODO: This is a placeholder implementation that just counts messages per user
        Should be enhanced later with proper summarization logic
        """
        # Get messages from last 24 hours
        twenty_four_hours_ago = datetime.now() - timedelta(hours=24)
        message_counts = self.message_store.get_message_count_by_user(twenty_four_hours_ago)
        
        if not message_counts:
            return "No messages to summarize from the last 24 hours."
        
        # Generate placeholder summary
        summary_lines = ["📊 **Daily Channel Summary** 📊"]
        summary_lines.append(f"*Summary period: Last 24 hours*\n")
        
        # TODO: Replace this simple counting with actual summarization
        summary_lines.append("**Message activity by user:**")
        for author_name, count in message_counts.items():
            summary_lines.append(f"• {author_name}: {count} messages")
        
        # TODO: Add more sophisticated analysis here
        summary_lines.append("\n*This is a placeholder summary. TODO: Implement proper summarization logic*")
        
        return "\n".join(summary_lines)

    async def send_summary_to_channel(self, bot: commands.Bot, channel_id: int) -> bool:
        """Send the generated summary to the specified channel"""
        try:
            summary_content = self.generate_daily_summary()
            channel = bot.get_channel(channel_id)
            
            if isinstance(channel, TextChannel):
                await channel.send(summary_content)
                return True
            else:
                print(f"Channel {channel_id} is not a text channel")
                return False
                
        except Exception as e:
            print(f"Error sending summary: {e}")
            return False

    def get_summary_schedule(self) -> datetime:
        """Get the next scheduled summary time"""
        return self.config.get_summary_time()
