pyproject.toml
[project]
name = "summarizer-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "toml>=0.10.2",
]
